post

+++ app/Http/Controllers/TasksController.php
@@ -17,12 +17,38 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
         
+        //$tasks = Task::all();
+        
+
+        
+        
+        if (\Auth::check()) {
+            
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            return view('tasks.index', [
+            'tasks' => $tasks
+            ]);
+        }
+
+        else{
+            return view('welcome');
+        }
+        
+        
+    }
+    
+        
+/**
+ *         $tasks = Task::all();
         return view('tasks.index', [
             'tasks' => $tasks
             ]);
-    }
+            
+    
+*/
 
     /**
      * Show the form for creating a new resource.
@@ -52,11 +78,13 @@ class TasksController extends Controller
             
         ]);
 
-        
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::user()->id;
         $task->save();
+        
+        
 
         return redirect('/');
     }
@@ -70,11 +98,20 @@ class TasksController extends Controller
      
     public function show($id)
     {
+        
         $task = Task::find($id);
         
+        if(\Auth::user()->id === $task->user_id){
+       
         return view('tasks.show', [
             'task' => $task,
             ]);
+        }
+    
+
+        else {return redirect('/');
+    }
+    
     }
 
     /**
@@ -122,9 +159,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
+        $task = \App\Task::find($id);
+        
+        if(\Auth::user()->id === $task->user_id){
         $task->delete();
+        
+        }
 
-        return redirect('/');
+        return redirect('');
     }
 }
